+++ CHANGELOG.md
@@ -13,5 +13,6 @@ Changelog for the KiCad Partslist Editor
 * Changed config-version for FieldKeywords.conf from 1.0 to 1.1
 * remove duplicate listings of parts which are in multiple subsheet instances
 * fixed major bug, which limited the number of fields to 9 or less (parsing error)
-* changed separator for CSV from ',' to ';'
+* make CSV separator configurable with config.ini
+* using os module for file name and path manipulations